requirements.txt
faster-whisper
silero-vad
fastapi
uvicorn[standard]
websockets
sqlalchemy
pydantic
pyyaml
httpx
transformers
pytest
